Add limitations for CurrentUser Advice + Add function to retrieve user role
CurrentUserControllerAdvice allowed access to full user object, this has
now been limited to the allow only access to the UserDto in addition to that
a function was added to the UserService that allows the retrieval of a given
users role. It has been added to the CurrentUserControllerAdvice to allow
checks for the users role without giving them access to critical data like
passwords.